+++ src/js/api.ts
@@ -501,8 +501,19 @@ export class PyodideAPI {
       );
     }
 
-    if (Module.FS.findObject(path) == null) {
-      Module.FS.mkdirTree(path);
+    Module.FS.mkdirTree(path);
+    const { node } = Module.FS.lookupPath(path, {
+      follow_mount: false,
+    });
+
+    if (FS.isMountpoint(node)) {
+      throw new Error(`path '${path}' is already a file system mount point`);
+    }
+    if (!FS.isDir(node.mode)) {
+      throw new Error(`path '${path}' points to a file not a directory`);
+    }
+    for (const _ in node.contents) {
+      throw new Error(`directory '${path}' is not empty`);
     }
 
     Module.FS.mount(