+++ src/js/types.ts
@@ -217,6 +217,7 @@ export interface FS {
   stat: (path: string, dontFollow?: boolean) => any;
   readdir: (node: FSNode) => string[];
   isDir: (mode: number) => boolean;
+  isMountpoint: (mode: FSNode) => boolean;
   lookupPath: (path: string) => { node: FSNode };
   isFile: (mode: number) => boolean;
   writeFile: (path: string, contents: any, o?: { canOwn?: boolean }) => void;