+++ src/tests/test_filesystem.py
@@ -257,6 +257,63 @@ def test_nativefs_dir(request, selenium_standalone):
     )
 
 
+@only_chrome
+def test_nativefs_errors(selenium):
+    selenium.run_js(
+        """
+        const root = await navigator.storage.getDirectory();
+        const handle = await root.getDirectoryHandle("dir", { create: true });
+
+        await pyodide.mountNativeFS("/mnt1/nativefs", handle);
+        assertThrowsAsync(
+          async () => await pyodide.mountNativeFS("/mnt1/nativefs", handle),
+          "Error",
+          "path '/mnt1/nativefs' is already a file system mount point",
+        );
+
+        pyodide.FS.mkdirTree("/mnt2");
+        pyodide.FS.writeFile("/mnt2/some_file", "contents");
+        assertThrowsAsync(
+          async () => await pyodide.mountNativeFS("/mnt2/some_file", handle),
+          "Error",
+          "path '/mnt2/nativefs' points to a file not a directory",
+        );
+        // Check we didn't overwrite the file.
+        assert(
+          () =>
+            pyodide.FS.readFile("/mnt2/some_file", { encoding: "utf8" }) === "contents",
+        );
+
+        pyodide.FS.mkdirTree("/mnt3/nativefs");
+        pyodide.FS.writeFile("/mnt3/nativefs/a.txt", "contents");
+        assertThrowsAsync(
+          async () => await pyodide.mountNativeFS("/mnt3/nativefs", handle),
+          "Error",
+          "directory '/mnt3/nativefs' is not empty",
+        );
+        // Check directory wasn't changed
+        const { node } = pyodide.FS.lookupPath("/mnt3/nativefs/");
+        assert(() => Object.entries(node.contents).length === 1);
+        assert(
+          () =>
+            pyodide.FS.readFile("/mnt3/nativefs/a.txt", { encoding: "utf8" }) ===
+            "contents",
+        );
+
+        const [r1, r2] = await Promise.allSettled([
+          pyodide.mountNativeFS("/mnt4/nativefs", handle),
+          pyodide.mountNativeFS("/mnt4/nativefs", handle),
+        ]);
+        assert(() => r1.status === "fulfilled");
+        assert(() => r2.status === "rejected");
+        assert(
+          () =>
+            r2.reason.message === "path '/mnt4/nativefs' is already a file system mount point",
+        );
+        """
+    )
+
+
 @pytest.fixture
 def browser(selenium):
     return selenium.browser